Move navigation classes to PhpMyAdmin namespace
