+++ libraries/classes/Navigation/Nodes/NodeDatabaseChildContainer.php
@@ -5,7 +5,7 @@
  *
  * @package PhpMyAdmin-Navigation
  */
-namespace PMA\libraries\navigation\nodes;
+namespace PhpMyAdmin\Navigation\Nodes;
 
 /**
  * Represents container node that carries children of a database
